Separate a new SelfJoinEvent type out of Slack JoinEvent. (#49)
This new type exists solely for the benefit of the IRC server, which apparently needs to send NAMES and the channel topic before clients recognize that the channel has been joined. (HexChat will not display user lists until this has been performed, for example.)

We wire up the existing `handleChannelJoined` functionality (which was previously only executed when an IRC client sent a JOIN, or on IRC client connect) to also trigger when Slack sends a `member_joined_channel` event over the WebSocket connection.

Note that the similarly named `channel_joined` event is still unhandled, and is not necessary for proper functioning.